The act_informative has been separated to commands.
