+++ src/engine/ui/cmd/do_time.h
@@ -1,0 +1,17 @@
+/**
+\file do_time.h - a part of the Bylins engine.
+\authors Created by Sventovit.
+\date 21.09.2024.
+\brief Brief description.
+\detail Detail description.
+*/
+
+#ifndef BYLINS_SRC_ENGINE_UI_CMD_DO_TIME_H_
+#define BYLINS_SRC_ENGINE_UI_CMD_DO_TIME_H_
+
+class CharData;
+void do_time(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_ENGINE_UI_CMD_DO_TIME_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
