+++ src/gameplay/communication/check_invoice.h
@@ -1,0 +1,18 @@
+/**
+\file check_invoice.h - a part of the Bylins engine.
+\authors Created by Sventovit.
+\date 21.09.2024.
+\brief Brief description.
+\detail Detail description.
+*/
+
+#ifndef BYLINS_SRC_GAMEPLAY_COMMUNICATION_CHECK_INVOICE_H_
+#define BYLINS_SRC_GAMEPLAY_COMMUNICATION_CHECK_INVOICE_H_
+
+class CharData;
+bool login_change_invoice(CharData *ch);
+void god_work_invoice();
+
+#endif //BYLINS_SRC_GAMEPLAY_COMMUNICATION_CHECK_INVOICE_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
