+++ src/gameplay/mechanics/sight.h
@@ -28,6 +28,7 @@ const char *show_obj_to_char(ObjData *object, CharData *ch, int mode, int show_s
 void obj_info(CharData *ch, ObjData *obj, char buf[kMaxStringLength]);
 void print_zone_info(CharData *ch);
 const char *print_obj_state(int tm_pct);
+void diag_char_to_char(CharData *i, CharData *ch);
 
 #endif //BYLINS_SRC_GAME_MECHANICS_SIGHT_H_
